changed error.status, split controllers.contacts into files
