+++ middlewares/isValidId.js
@@ -1,12 +1,12 @@
-const { isValidObjectId } = require("mongoose");
-const { HttpError } = require("../helpers");
+const { isValidObjectId } = require('mongoose');
+const { HttpError } = require('../helpers');
 
 const isValidId = (req, res, next) => {
-  const { id } = req.params;
-  if (!isValidObjectId(id)) {
-    next(HttpError(400, `${id} is not a valid id`));
-  }
-  next();
+	const { id } = req.params;
+	if (!isValidObjectId(id)) {
+		next(HttpError(404, `${id} is not a valid id`));
+	}
+	next();
 };
 
 module.exports = isValidId;
